i#5767: Fix vs2019 pre-commit failures relating to zlib warnings (#5766)
Adds casts to satisfy vs2019 warnings-as-errors relating
to lossy implicit integer conversions.

Disables zlib for WIN32 AND AUTOMATED_TESTING to unblock the suite.
We should re-enable once we have proper zlib package identification
and installation on the GA CI machines.


Issue: #5767

+++ clients/drcachesim/common/gzip_ostream.h
@@ -79,7 +79,8 @@ public:
         }
         int res = traits_type::not_eof(extra_char);
         if (pptr() > pbase()) {
-            int len = gzwrite(file_, pbase(), pptr() - pbase());
+            int len =
+                gzwrite(file_, pbase(), static_cast<unsigned int>(pptr() - pbase()));
             if (len < pptr() - pbase())
                 res = traits_type::eof();
         }
